refactor: Move ggshield.iac to ggshield.verticals.iac
This is one of the step needed to regroup all verticals into one
package and simplify the identification of the verticals layer as
defined in #521.

refs: #521
