+++ ggshield/verticals/iac/collection/iac_path_scan_collection.py
@@ -2,7 +2,7 @@ from typing import Optional
 
 from pygitguardian.iac_models import IaCScanResult
 
-from ggshield.iac.collection.iac_scan_collection import (
+from ggshield.verticals.iac.collection.iac_scan_collection import (
     CollectionType,
     IaCScanCollection,
 )
