+++ ggshield/verticals/iac/output/iac_output_handler.py
@@ -4,9 +4,13 @@ from typing import Optional
 import click
 
 from ggshield.core.errors import ExitCode
-from ggshield.iac.collection.iac_diff_scan_collection import IaCDiffScanCollection
-from ggshield.iac.collection.iac_path_scan_collection import IaCPathScanCollection
-from ggshield.iac.collection.iac_scan_collection import (
+from ggshield.verticals.iac.collection.iac_diff_scan_collection import (
+    IaCDiffScanCollection,
+)
+from ggshield.verticals.iac.collection.iac_path_scan_collection import (
+    IaCPathScanCollection,
+)
+from ggshield.verticals.iac.collection.iac_scan_collection import (
     CollectionType,
     IaCScanCollection,
 )
