+++ tests/unit/verticals/iac/output/test_iac_json_output.py
