+++ tests/unit/verticals/iac/output/test_iac_text_output.py
